refactor store and static
