+++ src/components/projectPage/ProjectCanvas.js
@@ -1,8 +1,8 @@
 import React, {useEffect, useRef} from 'react';
 import styled from 'styled-components';
 import {useSelector} from 'react-redux';
-import {store} from '../../app/store';
-import {updateGroup} from '../../app/actions/index';
+import {store} from '../../app/store/index.js';
+import {updateGroup} from '../../app/actions/index.js';
 import * as PIXI from 'pixi.js';
 // import {SVGScene} from '@pixi-essentials/svg';
