+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import {initFirebase} from './app/utils/firebase.js';
-import {store} from './app/store';
+import {store} from './app/store/index.js';
 import {Provider} from 'react-redux';
 import * as serviceWorker from './serviceWorker';
